requirements.txt
streamlit>=1.30.0
python-pptx>=0.6.23
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.18.0
lxml>=4.9.0
openai>=1.10.0
